requirements.txt
streamlit
pandas
numpy
scikit-learn
requests
joblibstreamlit
joblib
python-whois
